requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
PyPDF2
pandas
tiktoken
langchain-google-genai